TestInterfaces.cs
This is the same file but with some notes. the constructors are objects, which using the parameters gives them functionality. The roundpan takes a parameter of a radius, so that whenever the roundpan constructor/object is called, the radius is defined with it; this allows for malleability and complexity without DRYing. Think through this... this is still solidifying in my mind, and this is a good example to see it in action. 

The Interfaces here only provide method names, but think through this as well. work in it with my own project. Let the usefulness come through discovery and experience. It's being able to bring separate 'elements' together by identifying their common logic, in this case, being the GetArea and GetCost. the interface allows them to be connected through these without creating more classes to support them. There is more to interfaces, but for the moment, this is pretty cool.